Create Naive_Bayes
Using scikit-learn
Now that we’ve formatted our data correctly, we can use it using scikit-learn’s MultinomialNB classifier.

This classifier can be trained using the .fit() method. .fit() takes two parameters: The array of data points (which we just made) and an array of labels corresponding to each data point.

Finally, once the model has been trained, we can use the .predict() method to predict the labels of new points. .predict() takes a list of points that you want to classify and it returns the predicted labels of those points.

Finally, .predict_proba() will return the probability of each label given a point. Instead of just returning whether the review was good or bad, it will return the likelihood of a good or bad review.

Note that in the code editor, we’ve imported some of the variables you created last time. Specifically, we’ve imported the counter object, training_counts and then make review_counts. This means the program won’t have to re-create those variables and should help the runtime of your program.