Create NumPy_SampleSizeDetermination
A/B Testing at Nosh Mish Mosh
The Nosh Mish Mosh is a recipe and ingredient meal delivery service. We ship the raw materials and you get to cook it at your home! We’ve decided to hire a data analyst to help them make product and interface decisions. Get started and we’ll be able to judge the amount of data we’ll need for these differences to be meaningful.

If you get stuck during this project or would like to see an experienced developer work through it, click “Get Help“ to see a project walkthrough video.